requirement.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
anyascii==0.3.3
anyio==4.11.0
asgiref==3.10.0
attrs==25.4.0
audioop-lts==0.2.2
audioread==3.1.0
babel==2.17.0
billiard==4.2.2
cachetools==6.2.1
celery==5.5.3
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
contourpy==1.3.3
coqpit-config==0.2.1
-e git+https://github.com/idiap/coqui-ai-TTS@c0749fb6b7e99638cfc12abc8fce7433309923a2#egg=coqui_tts  
coqui-tts-trainer==0.3.1
cron_descriptor==2.0.6
cryptography==46.0.3
cycler==0.12.1
Cython==3.1.6
dateparser==1.1.8
decorator==5.2.1
deprecation==2.1.0
Django==5.2.7
django-celery-beat==2.8.1
django-cors-headers==4.9.0
django-timezone-field==7.1
djangorestframework==3.16.1
docopt==0.6.2
einops==0.8.1
elevenlabs==2.20.1
encodec==0.1.1
filelock==3.20.0
fish-audio-sdk==2025.6.3
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.9.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.0
google-api-python-client==2.185.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.71.0
grpcio==1.76.0
grpcio-status==1.71.2
gruut==2.4.0
gruut-ipa==0.13.0
gruut_lang_de==2.0.1
gruut_lang_en==2.0.1
gruut_lang_es==2.0.1
gruut_lang_fr==2.0.2
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
httpx-ws==0.8.1
huggingface-hub==0.36.0
hyperframe==6.1.0
idna==3.11
inflect==7.5.0
Jinja2==3.1.6
joblib==1.5.2
jsonlines==1.2.0
kiwisolver==1.4.9
ko-speech-tools==0.1.0
kombu==5.5.4
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.45.1
Markdown==3.9
MarkupSafe==3.0.3
matplotlib==3.10.7
monotonic-alignment-search==0.2.1
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
networkx==3.5
num2words==0.5.14
numba==0.62.1
numpy==2.3.4
ormsgpack==1.11.0
packaging==25.0
pillow==12.0.0
platformdirs==4.5.0
pooch==1.8.2
postgrest==2.22.2
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.1.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
PyJWT==2.10.1
pyparsing==3.2.5
pysbd==0.3.4
python-crfsuite==0.9.11
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
realtime==2.22.2
redis==7.0.1
regex==2025.10.23
requests==2.32.5
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.3
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soxr==1.0.0
sqlparse==0.5.3
standard-aifc==3.13.0
standard-chunk==3.13.0
standard-sunau==3.13.0
storage3==2.22.2
StrEnum==0.4.15
supabase==2.22.2
supabase-auth==2.22.2
supabase-functions==2.22.2
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tokenizers==0.21.4
torch==2.8.0
torchaudio==2.8.0
tqdm==4.67.1
transformers==4.55.4
typeguard==4.4.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.22.0